实现图像识别超时中断机制

+++ src/main/java/priv/koishi/pmc/Bean/FindPositionConfig.java
@@ -37,7 +37,7 @@ public class FindPositionConfig {
     /**
      * 匹配失败后重试时间间隔
      */
-    long retryWait;
+    int retryWait;
 
     /**
      * 单次匹配最大时间
